backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
youtube-transcript-api>=0.6.0
openai>=1.50.0
httpx>=0.27.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
ruff>=0.7.0
mypy>=1.11.0
